Add in tests for single ranges
